new: periodic db update
